Better country error handling and refactor
- Allow "root" landing page
- Alert user if any of the selected countries is missing data
- Refactor pure data
- Refactor some code in App.jsx
